Rename lib/auth in preparation for authorization related code
