+++ core/app/(user)/login/LoginForm.tsx
@@ -12,7 +12,7 @@ import { Form, FormControl, FormField, FormItem, FormLabel } from "ui/form";
 import { Check, Loader2 } from "ui/icon";
 import { Input } from "ui/input";
 
-import * as actions from "~/lib/auth/actions";
+import * as actions from "~/lib/authentication/actions";
 import { useServerAction } from "~/lib/serverActions";
 
 export const loginFormSchema = z.object({
