+++ core/app/(user)/login/page.tsx
@@ -2,7 +2,7 @@ import { headers } from "next/headers";
 import Link from "next/link";
 import { redirect } from "next/navigation";
 
-import { getLoginData } from "~/lib/auth/loginData";
+import { getLoginData } from "~/lib/authentication/loginData";
 import LoginForm from "./LoginForm";
 import { Notice } from "./Notice";
