+++ core/app/c/[communitySlug]/forms/[formSlug]/edit/actions.ts
@@ -4,7 +4,7 @@ import type { FormsId } from "db/public";
 
 import { db } from "~/kysely/database";
 import { isUniqueConstraintError } from "~/kysely/errors";
-import { getLoginData } from "~/lib/auth/loginData";
+import { getLoginData } from "~/lib/authentication/loginData";
 import { logError } from "~/lib/logging";
 import { autoRevalidate } from "~/lib/server/cache/autoRevalidate";
 import { defineServerAction } from "~/lib/server/defineServerAction";
