+++ core/app/c/[communitySlug]/members/[[...add]]/actions.ts
@@ -8,8 +8,8 @@ import { MemberRole } from "db/public";
 import type { TableMember } from "./getMemberTableColumns";
 import type { UserWithMember } from "~/lib/types";
 import { db } from "~/kysely/database";
-import { getLoginData } from "~/lib/auth/loginData";
-import { isCommunityAdmin as isAdminOfCommunity } from "~/lib/auth/roles";
+import { getLoginData } from "~/lib/authentication/loginData";
+import { isCommunityAdmin as isAdminOfCommunity } from "~/lib/authentication/roles";
 import { findCommunityBySlug } from "~/lib/server/community";
 import { defineServerAction } from "~/lib/server/defineServerAction";
 import * as Email from "~/lib/server/email";
