+++ core/app/c/[communitySlug]/stages/manage/components/panel/StagePanelActions.tsx
@@ -6,7 +6,7 @@ import { Card, CardContent } from "ui/card";
 import type { PageContext } from "~/app/components/ActionUI/PubsRunActionDropDownMenu";
 import { ActionConfigFormWrapper } from "~/app/components/ActionUI/ActionConfigFormWrapper";
 import { SkeletonCard } from "~/app/components/skeletons/SkeletonCard";
-import { getLoginData } from "~/lib/auth/loginData";
+import { getLoginData } from "~/lib/authentication/loginData";
 import { getStage, getStageActions } from "~/lib/db/queries";
 import { addAction, deleteAction } from "../../actions";
 import { StagePanelActionCreator } from "./StagePanelActionCreator";
