+++ core/app/c/[communitySlug]/stages/page.tsx
@@ -6,7 +6,7 @@ import type { UsersId } from "db/public";
 import { AuthTokenType } from "db/public";
 
 import { CreatePubButton } from "~/app/components/pubs/CreatePubButton";
-import { getPageLoginData } from "~/lib/auth/loginData";
+import { getPageLoginData } from "~/lib/authentication/loginData";
 import { findCommunityBySlug } from "~/lib/server/community";
 import { createToken } from "~/lib/server/token";
 import { StageList } from "./components/StageList";
