+++ core/lib/authentication/cookies.ts
