+++ core/lib/authentication/createMagicLink.ts
