+++ core/lib/authentication/password.ts
